config: add function get language from the user's browser header
This will scan the locales directory for any translations available there. It will extract the 2 character language code between "locale-" and ".php". It will then scan for the user's preferred language based on the header. If there isn't a translation for this, default to the user's defined config language. 

inspiration taken from stackoverflow for this, modified for more dynamic code.